% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.sync-package.R
\docType{package}
\name{settings.sync-package}
\alias{settings.sync}
\alias{settings.sync-package}
\title{settings.sync: RStudio Settings Sync}
\description{
RStudio addin that allows settings to be synced to GitHub Gists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjpalanca/settings.sync}
  \item Report bugs at \url{https://github.com/tjpalanca/settings.sync/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
