% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_addin}
\alias{sync_addin}
\alias{sync_addin_ui}
\alias{sync_addin_server}
\alias{sync_gist_id}
\alias{sync_gist_id_reset}
\alias{sync_config_dir}
\alias{sync_new}
\alias{sync_existing}
\title{Sync Addin}
\usage{
sync_addin()

sync_addin_ui(request)

sync_addin_server(input, output, session)

sync_gist_id(gist_id = NULL)

sync_gist_id_reset()

sync_config_dir()

sync_new()

sync_existing(gist_id = NULL)
}
\description{
This Gadget syncs rstudio settings to a GitHub gist.
}
\section{Functions}{
\itemize{
\item \code{sync_addin_ui}: Actual Sync Addin

\item \code{sync_addin_server}: Server Function

\item \code{sync_gist_id}: Gist ID checker
If no value is supplied, then retrieves the stored gist ID. If a value is
supplied, then writes the Gist ID into a text file defined by the user
directory.

\item \code{sync_gist_id_reset}: Reset the Gist ID

\item \code{sync_config_dir}: RStudio Config Directory

\item \code{sync_new}: Sync a new gist
This is used by the addin if there is no gist ID specified. It creates a new
gist and uploads the files into the gist. It does this in a weird way
(one-by-one) because sometimes the gist contents are truncated.

\item \code{sync_existing}: Sync and existing Gist
This is used when there is an existing configuraiton. It uses the last
modified timestamp of each file to determine whether it is more or less
updated than the gist file, and then syncs those changes.
}}

